# frontend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Document Archive API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/mindmap/tree": {
      "get": {
        "summary": "Get Mindmap Tree",
        "operationId": "get_mindmap_tree_api_mindmap_tree_get",
        "parameters": [
          {
            "name": "category_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "선택된 카테고리명",
              "title": "Category Name"
            },
            "description": "선택된 카테고리명"
          },
          {
            "name": "tag_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "선택된 태그명",
              "title": "Tag Name"
            },
            "description": "선택된 태그명"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "제목/설명 키워드",
              "title": "Q"
            },
            "description": "제목/설명 키워드"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          },
          {
            "name": "tag_limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 120,
              "title": "Tag Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MindMapTreeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MindMapCategoryNode": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "document_count": {
            "type": "integer",
            "title": "Document Count"
          },
          "latest_event_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Event Date"
          }
        },
        "type": "object",
        "required": [
          "category",
          "document_count"
        ],
        "title": "MindMapCategoryNode"
      },
      "MindMapDocumentNode": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "event_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Date"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "file_count": {
            "type": "integer",
            "title": "File Count",
            "default": 0
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "category",
          "updated_at"
        ],
        "title": "MindMapDocumentNode"
      },
      "MindMapTagNode": {
        "properties": {
          "tag": {
            "type": "string",
            "title": "Tag"
          },
          "document_count": {
            "type": "integer",
            "title": "Document Count"
          },
          "latest_event_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Event Date"
          }
        },
        "type": "object",
        "required": [
          "tag",
          "document_count"
        ],
        "title": "MindMapTagNode"
      },
      "MindMapTreeResponse": {
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "selected_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Category"
          },
          "selected_tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Tag"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/MindMapCategoryNode"
            },
            "type": "array",
            "title": "Categories"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/MindMapTagNode"
            },
            "type": "array",
            "title": "Tags"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/MindMapDocumentNode"
            },
            "type": "array",
            "title": "Documents"
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "default": 1
          },
          "size": {
            "type": "integer",
            "title": "Size",
            "default": 20
          },
          "total_documents": {
            "type": "integer",
            "title": "Total Documents",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "generated_at"
        ],
        "title": "MindMapTreeResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}